Move server-specific tables to subpackages (yay, organization)
