+++ loritta-discord/src/main/java/com/mrpowergamerbr/loritta/website/views/subviews/api/config/types/AutorolePayload.kt
@@ -9,7 +9,7 @@ import com.mrpowergamerbr.loritta.dao.ServerConfig
 import com.mrpowergamerbr.loritta.network.Databases
 import com.mrpowergamerbr.loritta.userdata.MongoServerConfig
 import net.dv8tion.jda.api.entities.Guild
-import net.perfectdreams.loritta.dao.AutoroleConfig
+import net.perfectdreams.loritta.dao.servers.moduleconfigs.AutoroleConfig
 import net.perfectdreams.loritta.website.session.LorittaJsonWebSession
 import org.jetbrains.exposed.sql.transactions.transaction
