+++ loritta-discord/src/main/java/net/perfectdreams/loritta/dao/servers/moduleconfigs/AutoroleConfig.kt
@@ -1,6 +1,6 @@
-package net.perfectdreams.loritta.dao
+package net.perfectdreams.loritta.dao.servers.moduleconfigs
 
-import net.perfectdreams.loritta.tables.AutoroleConfigs
+import net.perfectdreams.loritta.tables.servers.moduleconfigs.AutoroleConfigs
 import org.jetbrains.exposed.dao.Entity
 import org.jetbrains.exposed.dao.EntityClass
 import org.jetbrains.exposed.dao.EntityID
