+++ loritta-discord/src/main/java/net/perfectdreams/loritta/dao/servers/moduleconfigs/EconomyConfig.kt
@@ -1,6 +1,6 @@
-package net.perfectdreams.loritta.dao
+package net.perfectdreams.loritta.dao.servers.moduleconfigs
 
-import net.perfectdreams.loritta.tables.EconomyConfigs
+import net.perfectdreams.loritta.tables.servers.moduleconfigs.EconomyConfigs
 import org.jetbrains.exposed.dao.Entity
 import org.jetbrains.exposed.dao.EntityClass
 import org.jetbrains.exposed.dao.EntityID
