+++ loritta-discord/src/main/java/net/perfectdreams/loritta/dao/servers/moduleconfigs/EventLogConfig.kt
@@ -1,6 +1,6 @@
-package net.perfectdreams.loritta.dao
+package net.perfectdreams.loritta.dao.servers.moduleconfigs
 
-import net.perfectdreams.loritta.tables.EventLogConfigs
+import net.perfectdreams.loritta.tables.servers.moduleconfigs.EventLogConfigs
 import org.jetbrains.exposed.dao.Entity
 import org.jetbrains.exposed.dao.EntityClass
 import org.jetbrains.exposed.dao.EntityID
