+++ loritta-discord/src/main/java/net/perfectdreams/loritta/tables/servers/moduleconfigs/LevelConfigs.kt
@@ -1,4 +1,4 @@
-package net.perfectdreams.loritta.tables
+package net.perfectdreams.loritta.tables.servers.moduleconfigs
 
 import com.mrpowergamerbr.loritta.utils.exposed.array
 import net.perfectdreams.loritta.utils.levels.RoleGiveType
