+++ loritta-discord/src/main/java/net/perfectdreams/loritta/website/utils/config/types/TwitchConfigTransformer.kt
@@ -7,10 +7,7 @@ import com.google.gson.JsonObject
 import com.mrpowergamerbr.loritta.dao.ServerConfig
 import com.mrpowergamerbr.loritta.network.Databases
 import net.dv8tion.jda.api.entities.Guild
-import net.perfectdreams.loritta.tables.TrackedTwitchAccounts
-import net.perfectdreams.loritta.tables.TrackedYouTubeAccounts
-import net.perfectdreams.loritta.website.session.LorittaJsonWebSession
-import org.jetbrains.exposed.sql.SqlExpressionBuilder.eq
+import net.perfectdreams.loritta.tables.servers.moduleconfigs.*
 import org.jetbrains.exposed.sql.deleteWhere
 import org.jetbrains.exposed.sql.insert
 import org.jetbrains.exposed.sql.select
